backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose[cryptography]
passlib[argon2]
python-multipart
